refactor: state로 상태 정보 넘겨주는 동작 방지 및 ContextProcider 위치 변경
